fix name interface
