+++ src/main/java/com/example/AMSProject/controller/AdvertisementController.java
@@ -2,7 +2,6 @@ package com.example.AMSProject.controller;
 
 import com.example.AMSProject.model.*;
 import com.example.AMSProject.service.AdvertisementService;
-import com.example.AMSProject.service.AdvertisementServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,7 +14,7 @@ import java.text.ParseException;
 public class AdvertisementController {
 
    @Autowired
-   private AdvertisementServiceImpl amsService;
+   private AdvertisementService amsService;
 
     /**
      * POST /advertisement/setViewed
